+++ NINA.WPF.Base/ViewModel/Equipment/FilterWheel/FilterWheelChooserVM.cs
@@ -26,6 +26,7 @@ using NINA.Core.Locale;
 using NINA.Equipment.Interfaces;
 using NINA.Equipment.Equipment;
 using NINA.Equipment.Equipment.MyCamera;
+using NINA.Equipment.Equipment.MyCamera.ToupTekAlike;
 using NINA.Equipment.SDK.CameraSDKs.SBIGSDK;
 using System.Threading.Tasks;
 using NINA.Equipment.Interfaces.ViewModel;
@@ -147,6 +148,111 @@ namespace NINA.WPF.Base.ViewModel.Equipment.FilterWheel {
                         var fw = new PlayerOneFilterWheel(i, profileService);
                         Logger.Debug($"Adding PlayerOne Filter Wheel {i})");
                         devices.Add(fw);
+                    }
+                } catch (Exception ex) {
+                    Logger.Error(ex);
+                }
+
+                /* Altair */
+                try {
+                    var altairDevices = Altair.Altaircam.EnumV2();
+                    Logger.Info($"Found {altairDevices?.Length} Altair Devices");
+                    foreach (var instance in altairDevices) {
+                        var info = instance.ToDeviceInfo();
+                        if (((ToupTekAlikeFlag)info.model.flag & ToupTekAlikeFlag.FLAG_FILTERWHEEL) > 0) {
+                            var wheel = new ToupTekAlikeFilterWheel(info, new AltairSDKWrapper(), profileService);
+                            devices.Add(wheel);
+                        }
+                    }
+                } catch (Exception ex) {
+                    Logger.Error(ex);
+                }
+
+                /* ToupTek */
+                try {
+                    var toupTekDevices = ToupTek.ToupCam.EnumV2();
+                    Logger.Info($"Found {toupTekDevices?.Length} ToupTek Devices");
+                    foreach (var instance in toupTekDevices) {
+                        var info = instance.ToDeviceInfo();
+                        if (((ToupTekAlikeFlag)info.model.flag & ToupTekAlikeFlag.FLAG_FILTERWHEEL) > 0) {
+                            var wheel = new ToupTekAlikeFilterWheel(info, new ToupTekSDKWrapper(), profileService);
+                            devices.Add(wheel);
+                        }
+                    }
+                } catch (Exception ex) {
+                    Logger.Error(ex);
+                }
+
+                /* Ogma */
+                try {
+                    var ogmaDevices = Ogmacam.EnumV2();
+                    Logger.Info($"Found {ogmaDevices?.Length} Ogma Devices");
+                    foreach (var instance in ogmaDevices) {
+                        var info = instance.ToDeviceInfo();
+                        if (((ToupTekAlikeFlag)info.model.flag & ToupTekAlikeFlag.FLAG_FILTERWHEEL) > 0) {
+                            var wheel = new ToupTekAlikeFilterWheel(info, new OgmaSDKWrapper(), profileService);
+                            devices.Add(wheel);
+                        }
+                    }
+                } catch (Exception ex) {
+                    Logger.Error(ex);
+                }
+
+                /* Omegon */
+                try {
+                    var omegonDevices = Omegon.Omegonprocam.EnumV2();
+                    Logger.Info($"Found {omegonDevices?.Length} Omegon Devices");
+                    foreach (var instance in omegonDevices) {
+                        var info = instance.ToDeviceInfo();
+                        if (((ToupTekAlikeFlag)info.model.flag & ToupTekAlikeFlag.FLAG_FILTERWHEEL) > 0) {
+                            var wheel = new ToupTekAlikeFilterWheel(info, new OmegonSDKWrapper(), profileService);
+                            devices.Add(wheel);
+                        }
+                    }
+                } catch (Exception ex) {
+                    Logger.Error(ex);
+                }
+
+                /* Risingcam */
+                try {
+                    var risingCamDevices = Nncam.EnumV2();
+                    Logger.Info($"Found {risingCamDevices?.Length} RisingCam Devices");
+                    foreach (var instance in risingCamDevices) {
+                        var info = instance.ToDeviceInfo();
+                        if (((ToupTekAlikeFlag)info.model.flag & ToupTekAlikeFlag.FLAG_FILTERWHEEL) > 0) {
+                            var wheel = new ToupTekAlikeFilterWheel(info, new RisingcamSDKWrapper(), profileService);
+                            devices.Add(wheel);
+                        }
+                    }
+                } catch (Exception ex) {
+                    Logger.Error(ex);
+                }
+
+                /* MallinCam */
+                try {
+                    var mallinCamDevices = MallinCam.Mallincam.EnumV2();
+                    Logger.Info($"Found {mallinCamDevices?.Length} MallinCam Devices");
+                    foreach (var instance in mallinCamDevices) {
+                        var info = instance.ToDeviceInfo();
+                        if (((ToupTekAlikeFlag)info.model.flag & ToupTekAlikeFlag.FLAG_FILTERWHEEL) > 0) {
+                            var wheel = new ToupTekAlikeFilterWheel(info, new MallinCamSDKWrapper(), profileService);
+                            devices.Add(wheel);
+                        }
+                    }
+                } catch (Exception ex) {
+                    Logger.Error(ex);
+                }
+
+                /* SVBony */
+                try {
+                    var svBonyDevices = Svbonycam.EnumV2();
+                    Logger.Info($"Found {svBonyDevices?.Length} SVBony Devices");
+                    foreach (var instance in svBonyDevices) {
+                        var info = instance.ToDeviceInfo();
+                        if (((ToupTekAlikeFlag)info.model.flag & ToupTekAlikeFlag.FLAG_FILTERWHEEL) > 0) {
+                            var wheel = new ToupTekAlikeFilterWheel(info, new SVBonySDKWrapper(), profileService);
+                            devices.Add(wheel);
+                        }
                     }
                 } catch (Exception ex) {
                     Logger.Error(ex);