+++ NINA/View/Equipment/FilterWheel/FilterWheelTemplateSelector.cs
@@ -23,6 +23,7 @@ namespace NINA.View.Equipment {
     internal class FilterWheelTemplateSelector : DataTemplateSelector {
         public DataTemplate Default { get; set; }
         public DataTemplate Zwo { get; set; }
+        public DataTemplate ToupTekAlike { get; set; }
         public DataTemplate FailedToLoadTemplate { get; set; }
 
         public string Postfix { get; set; }
@@ -30,6 +31,8 @@ namespace NINA.View.Equipment {
         public override DataTemplate SelectTemplate(object item, DependencyObject container) {
             if (item is ASIFilterWheel) {
                 return Zwo;
+            } else if (item is ToupTekAlikeFilterWheel) {
+                return ToupTekAlike;
             } else {
                 var templateKey = item?.GetType().FullName + Postfix;
                 if (item != null && Application.Current.Resources.Contains(templateKey)) {