+++ RELEASE_NOTES.md
@@ -19,6 +19,7 @@ More details at <a href="https://nighttime-imaging.eu/donate/" target="_blank">n
 - Sequencer conditions and instructions now better handle transitions near the horizon, preventing sudden jumps in sun and moon altitude caused by atmospheric refraction effects breaking down below horizon.
 - Direct guider will now wait for "IsPulseGuiding" flag to become false before continuing.
 - In Legacy Sequencer, when interrupting the start or end actions, the sequence will now properly re-run these when starting again
+- ToupTek based filter wheels and focusers will no longer be listed in the camera connector.
  
 ### **Device Management**
 - **Device Chooser Enhancements**  
@@ -36,6 +37,8 @@ More details at <a href="https://nighttime-imaging.eu/donate/" target="_blank">n
     - Optionally disallow the mount to be unparked if the dome or roof controller reports a shutter state other than Open.
 - ** Switch Polling**
     - Switches are now polled sequentially instead of in parallel for their status update, to accommodate drivers that do not handle concurrent access properly.
+- **Altair, Mallincam, Ogma, Omegon, Risingcam, SvBony and ToupTek Filterwheel Native Driver**
+  - The ToupTek based filter wheels are now available as a native driver.
 
 ### **User Interface & Usability**
 - **Framing Assistant Improvements**